+++ src/main/java/org/opentripplanner/standalone/config/routerconfig/VectorTileConfig.java
@@ -6,20 +6,22 @@ import static org.opentripplanner.inspector.vector.LayerParameters.MAX_ZOOM;
 import static org.opentripplanner.inspector.vector.LayerParameters.MIN_ZOOM;
 import static org.opentripplanner.standalone.config.framework.json.OtpVersion.V2_0;
 import static org.opentripplanner.standalone.config.framework.json.OtpVersion.V2_5;
+import static org.opentripplanner.standalone.config.framework.json.OtpVersion.V2_6;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import javax.annotation.Nullable;
 import org.opentripplanner.ext.vectortiles.VectorTilesResource;
+import org.opentripplanner.ext.vectortiles.VectorTilesResource.LayerType;
+import org.opentripplanner.ext.vectortiles.layers.LayerFilters;
 import org.opentripplanner.inspector.vector.LayerParameters;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 
-public class VectorTileConfig
-  implements VectorTilesResource.LayersParameters<VectorTilesResource.LayerType> {
+public class VectorTileConfig implements VectorTilesResource.LayersParameters<LayerType> {
 
   public static final VectorTileConfig DEFAULT = new VectorTileConfig(List.of(), null, null);
-  private final List<LayerParameters<VectorTilesResource.LayerType>> layers;
+  private final List<LayerParameters<LayerType>> layers;
 
   @Nullable
   private final String basePath;
@@ -28,7 +30,7 @@ public class VectorTileConfig
   private final String attribution;
 
   VectorTileConfig(
-    Collection<? extends LayerParameters<VectorTilesResource.LayerType>> layers,
+    Collection<? extends LayerParameters<LayerType>> layers,
     @Nullable String basePath,
     @Nullable String attribution
   ) {
@@ -38,7 +40,7 @@ public class VectorTileConfig
   }
 
   @Override
-  public List<LayerParameters<VectorTilesResource.LayerType>> layers() {
+  public List<LayerParameters<LayerType>> layers() {
     return layers;
   }
 
@@ -144,7 +146,18 @@ public class VectorTileConfig
           "The value is a fraction of the tile size. If you are having problem with icons and " +
           "shapes being clipped at tile edges, then increase this number."
         )
-        .asDouble(EXPANSION_FACTOR)
+        .asDouble(EXPANSION_FACTOR),
+      node
+        .of("filter")
+        .since(V2_6)
+        .summary("Reduce the result set of a layer further by a specific filter.")
+        .description(
+          """
+          This is useful for when the schema of a layer, say stops, should remain unchanged but some
+          elements should not be included in the result.
+          """
+        )
+        .asEnum(LayerFilters.FilterType.NONE)
     );
   }
 
@@ -155,7 +168,8 @@ public class VectorTileConfig
     int maxZoom,
     int minZoom,
     int cacheMaxSeconds,
-    double expansionFactor
+    double expansionFactor,
+    LayerFilters.FilterType filterType
   )
     implements LayerParameters<VectorTilesResource.LayerType> {}
 }